Update index.html
I haven't previewed this, but I noticed a second <body> tag with a 'background' attribute. Removing that and moving the attribute up to the first <body> tag is probably the right thing to do. Apologies, this was a very quick edit so if this doesn't work, feel free to shout at me.